requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
av==12.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
cachetools==5.3.3
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comtypes==1.4.4
contourpy==1.2.1
cryptography==42.0.8
cssselect2==0.7.0
ctranslate2==4.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
easyocr==1.7.1
einops==0.8.0
et-xmlfile==1.1.0
executing==2.0.1
faster-whisper==1.0.2
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
googlesearch-python==1.2.4
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
HyperPyYAML==1.2.2
icalendar==5.0.13
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
intel-openmp==2021.4.0
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.2
lightning-utilities==0.11.3.post0
llvmlite==0.43.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mouse==0.7.1
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
noisereduce==3.0.2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.0
omegaconf==2.3.0
onnxruntime==1.18.1
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.82
openpyxl==3.1.5
optuna==3.6.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
playwright==1.44.0
pooch==1.8.2
primePy==1.3
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20240210
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyee==11.1.0
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
pyloudnorm==0.1.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.7
pyotp==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-nmap==0.7.1
python-pptx==0.6.23
pytorch-lightning==2.3.2
pytorch-metric-learning==2.5.0
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.21.0
semver==3.0.2
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SpeechRecognition==3.10.4
speedtest-cli==2.1.3
SQLAlchemy==2.0.31
srt==3.5.3
stack-data==0.6.3
svg.path==6.3
svglib==1.5.1
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tensorboardX==2.6.2.2
textblob==0.18.0.post0
threadpoolctl==3.5.0
tifffile==2024.5.22
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
tweepy==4.14.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualbox==2.1.1
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
whisper==1.1.10
win10toast==0.9
WMI==1.5.1
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
yfinance==0.2.40
